Improvements to multi objective (#52)
* Fix compiler SEGFAULT when file does not exist

* Fix superfluous '}' in CMakeLists

* make CBoundary copyable

* Made CRandom copyable while preserving semantic

* Made 2x2 crossover copyable while preserving semantic

* Added relative offspring size as command line parameter + fix templates

* Fix being unable to disable generation of .plot file

* Added stats printing to MO

* Added serialization to algorithms and individuals

* updated MO templates for serialization

* Added DE example

* Added missing required files

* Parallel stats calculation

* Removed warning

* Fix error on MacOS

* Maybe fixed Boost link with msvc

* Better support for old OpenMP versions (msvc)

* Fixed sinus_it DE for windows

* Fix unused include for windows compatibility

* Fix OpenMp errors on windows

* Added --verbose flag to build examples scripts

* Moved to C++17

* Generated Makefile is less verbose by default and Debug now implies -O0

* Fix UB on non updated templates

* Replaced deprecated STL functions

* Fix version used in verbose mode

* Removed old templates

* Better output of --help

* Fix --inputFile and --outputFile not being used + improved CIndividual
serialization.

* Added generic serialization to network

* Introducing CAlgorithmWrapper to deflate work done in templates

* Moved m_algorithm initialization to
ParametersImpl::newEvolutionnaryAlgorithm
+ cleaning

* Evaluation of initial population is now deferred at the first "run()".

* Fully switched to CMake + removed some useless code

* Fix runBreeding not working as intended

* restore old first generation m_iNbGen = 0

* Fix MOEAD

* Improved examples building script

* Added --silentNetwork option

* Fix invalid number of variables in TProblem

* Fix memory leak

* Fix 0 computed evaluations shown in CUDA template

* Fix two distinct buffer oveflow in Cnsga-iii

* Fix number of evaluations shown in statistics

* Better examples building script with `--ignore-errors`

* Fix memory leaks in CUDA_GP

* Fix SIGABRT because thread was not detached

* Fix listsort parameters containing absolute values instead of relative
ones.

* Default reduction parameter is now relative and equal to 100%

* Better CLI for compiler + fix CMAES CUDA not being used since the
beginning.

* Added option --examples-directory to script for building examples

* Fix memory leaks in CUDA_GP and CUDA + improved safety if an error
occurs and 0 cards are detected

* Fix buffer overflow by one

* Fix use after free in std::thread of CComUDPLayer if detached thread
survives its CComContext

* Ignore execution errors of CUDA binaries in CI because it is expected.

* Fix network and MO (crowding distance need recalculation + mutStep need
to be sent)

* Removed CMAES_CUDA template

* Added one example per template

* Added detection of UB, leaks and other errors in CI

* Added missing sudo to CI

* Added libomp-dev for clang CI

* Added network tests and load .pop tests

* fixed nsgaiii assert bug

* Fix network for templates using CEvolutionnaryAlgorithm

* Updated compiler message on how to compile

* fixed assert bug in DE.tpl

* Fix errors outside user code pointing to bogus line in .ez file

* Fix #22
EZ_PATH no longer requires a trailing '/' or '\'

* Better compiler errors 1/2

* Separated declarations from definitions inside lexer

* fix #30
Way better compiler errors

* Fix type error on MacOS

* replacemnt of m_generator from *.ez to *.tpl

* m_generated is deleted from *.ez in examles/templates/

* Moved tests if needed evaluation to CIndividual instead of templates.
+ All SO templates support forced reevaluation as a consequence

* aritty -> arity + removed duplicate code

* Removed duplicate getFitness

* Ensure crossover, printOn, clone, mutate, boundChecking, serialize, deserialize are virtual and overriding parent function

* Removed static CIndividual random generator unused everywhere

* Factored operator<< into .cpp

* arrity -> arity

* No more 'return fitness =' + allow lambdas in evaluator

* fix #18
Removed compiler action on 'random' and 'tossCoin' by adding function to
CRandomGenerator.h

* fixed assert in gde

* GDE.tpl is adapted to the AlgorithmWraper modification

* For GDE.tpl test : NB_OBJECTIVES is added in sinus_it.ez

* dtlz1 test is added for GDE

* gde.tpl: some warnings are deleted

* CDAS.tpl is fixed after refactoring

* Switch random engine to LCG + custom distribution. This commit brings
10x speedup on random number generation. fix #12

* Fix bounds of fast_bounded_distribution to match std:: ones + constexpr

* Added an assert if bias above 1

* Made .pf files be considered as binary

* Temporary remove for applying attributes

* PF files are now considered as binary

* Ignore bin files in github diff ?

* Better gitattributes ?

* Maybe ?

* Ignore pf in github diff ?

* CUDA codes is expected to crash on Linux CI (no GPU)

* Only ignore --ignore-errors executions, build should succeed

* Fix OpenMP msvc compatibility (int)

* Fixed small bug in CRandomGenerator - assert in tossCoin and better conversion from [) to []

* log in IBEA.tpl

* fixed assert in tossCoin

* Moved from cxxopts to boost::program_options + better version message
(include build type and compiler)

* Fix error in .prm generation that was there since the beginning

* Better errors and unified messages in compiler

* Fix short CLI option interfering with long ones after migration to
program_options

* Fix boost::multiprecision include blocking nvcc compilation

* Fix Boost link errors on msvc

* Added xoshiro256++

* Massive improvements to random number generation (x10 speedup). fix #12

* Fix warnings in Ccdas

* Fix operator-> on non clang

* Fix compatibility with very (very) old codes...

* MOEAD is fixed (problem of pop size definition)

* CDAS: fixed island model problem - added recalculation of converted objectives in case of island model when one individual is recieved

* MOEAD: fixed island model option (updates of neihbors with new solution)

* examples/templates/CDAS: fixed wrong path in pf

* Fix CLI parameters

* Definitely removed cxxopts

* Update flags for boolean options, while maintaining compatibility

* Added CAlgorithmWrapper to DE

* Cleaned up templates

---------

+++ libeasea/include/operators/selection/totalSelection.h
@@ -35,10 +35,10 @@ namespace selection
 template <typename TI, typename TIter>
 TIter totalSelection(std::vector<TI> &donorPop, TIter recipPopBegin, TIter recipPopEnd)
 {
-        if (donorPop.size() > std::distance(recipPopBegin, recipPopEnd)) 	LOG_ERROR(errorCode::value, "Donor populaton size must be smaller or the same as the recipient population!");
+        if (static_cast<long int>(donorPop.size()) > std::distance(recipPopBegin, recipPopEnd)) 	LOG_ERROR(errorCode::value, "Donor populaton size must be smaller or the same as the recipient population!");
 
         TIter selected = recipPopBegin;
-		if (std::distance(selected, recipPopEnd) >= donorPop.size())
+		if (std::distance(selected, recipPopEnd) >= static_cast<long int>(donorPop.size()))
 		{
 			for (size_t i = 0; i < donorPop.size(); ++i, ++selected)
                     	    *selected = *donorPop[i];
